"Refactor and clean up: Move `PickOne` struct from `config.h` to `utilpvt.h` to improve modularity, initialize class members directly in the `Config` constructor for consistent defaults, and remove redundant member initializations in `server.h`."
