+++ certs/authn/krb/configkrb.cpp
@@ -5,6 +5,7 @@
  */
 
 #include "configkrb.h"
+#include "utilpvt.h"
 
 namespace pvxs {
 namespace certs {
